// packages/sdk/src/client/api.ts
export interface Database {
    /**
     * Database\'s relative path in its package directory.
     * @type {string}
     * @memberof Database
     */
    'path'?: string;
    /**
     * Size of the embedded database in bytes.
     * @type {number}
     * @memberof Database
     */
    'size'?: number;
}